Rename (currently unused) config param 'peer.discovery' to 'peer.discovery.enabled'
Under config schemes like typesafe-config that let one treat config paths
like a directory structure, it is a problem to have a config name be both
a "leaf" with a value directly and a node or directly that contains subpaths.

Since 'peer.discovery' has subkeys, it may be better to let values attach
only to subkeys, so that directory style config libraries can be used.