Refactorings in anticipation of a public API for config plug-ins
Made abstract methods of SystemProperties protected rather than package visible.
Modified names and behaviors of the abstract methods to return null rather than
default values if a key is not found. The base class SystemProperties provides
the default if necessary, but we want individual subclasses to report their failure
so that we can implement a subclass that searches a path of implementations until
it finds a non-null value or must provide a default.