Updated testing to deal with a now immutable configuration object.
Eliminated calls to CONFIG.setDataBaseDir(...) in tests, as
SystemProperty objects are now immutable.

Modified gradle build to run each test in a forked JVM and
override each test's 'ethereumj.database.dir' with a unique
value.

(The tests do seem reliably to "work", but there are parallelism
issues re Gradle's forking that leave the directory names not
quite deterministic, and such that I'm not quite certain they are
always guaranteed to be unique.)