Fix Java Security global state in test suite
Under some orderings of running the test suite, the test
`org.ethereum.crypto.ECKeyTest.testSunECRoundTrip`
can fail.

Git Bisect reveals this was introduced in commit d4491a5

The `beforeAll` method in ECIESTest inserts the Bouncy Castle Security Provider
into the provider list with the highest priority.
This high priority can cause problems for other Java security code. Specifically,
in `ECKeyTest.testSunECRoundTrip`, when `sun.security.ec.ECKeyPairGenerator`
is looking for algorithm parameters, it ends up in
`org.spongycastle.jcajce.provider.asymmetric.ec.AlgorithmParametersSpi`.

The code that was removed in commit d4491a5
_appended_ the Bouncy Castle provider, so the conditional in `ECIESTest.beforeAll`
would be false. But after d4491a5, the conditional is true.

The fix is simple: just remove the provider insertion in `ECIESTest.beforeAll` as this
is not required.