Add ECDSASignature toHex() method
Converts the signature to a hex representation. There is some discussion on how the v value should be treated.
v can currently be a value of 0, 1, 27, or 28 (00, 01, 1b, 1c, in hex respectively). This method normalises v to be either 00, or 01.
The reasoning for this is that geth seems to also output with values in this range.