Allow encode method to handle case of nondirty Node instantiated by a parsedRlp.
Create a new method makeRlp() that reconstructs and returns an rlp from a
parsedRlp LList object.

Add (parsedRlp != null) case to the if (!dirty) section of the encode method.
Note that the newly created rlp length may be >= 32, so it may need to be saved
and returned as a hash.  The last section of the else clause handles saving a
new rlp.  Move this section of code out of the else clause so it can handle a
new rlp from either the if or else clause.