Fix RC in ChannelManager: notifyDisconnect vs processNewPeers
It occurred quite rare but turned into a ghost peer connection which is actually dead,
but triggered in tx and blocks propagation threads,
thus messages were queued in memory with no chance for release.
ChannelManager#newPeers is a type of CopyOnWriteArrayList and when notifyDisconnect was called
peer had been simply removed from a copy of newPeers list
while processNewPeers execution added it into activePeers at the same time.